Change: Add static analysis to CI and tighten build/tooling for warnings and vendor code
* Add clang-tidy and cppcheck to GitHub Actions and run them as dedicated build targets
* Extend .clang-tidy with cognitive complexity settings and ensure compile_commands.json is generated
* Introduce ENABLE_WERROR option and apply strict warning flags to project sources while treating vendor includes as SYSTEM
* Fix duplicate PIO header generation by using a target-specific OUTPUT_DIR for the example target
* Clean up TX queue APIs (const correctness, safer drain behavior) and small formatting/logic nits found by analyzers
* Document static analysis workflow in README

+++ src/examples/w55_udp_echo/main.c
@@ -49,7 +49,7 @@ static void w5500_init_no_linkwait(void)
         {2, 2, 2, 2, 2, 2, 2, 2}  // RX buffers
     };
 
-    if (ctlwizchip(CW_INIT_WIZCHIP, (void *)memsize) == -1)
+    if (ctlwizchip(CW_INIT_WIZCHIP, (void*)memsize) == -1)
     {
         printf("W5500 init failed (CW_INIT_WIZCHIP)\r\n");
     }
@@ -58,13 +58,12 @@ static void w5500_init_no_linkwait(void)
 static void configure_network(void)
 {
     // Configure network settings
-    wiz_NetInfo net_info = {
-        .mac = {0x02, 0x08, 0xDC, 0x12, 0x34, 0x56}, // MAC address
-        .ip = {192, 168, 29, 11},                    // IP address
-        .sn = {255, 255, 255, 0},                    // Subnet mask
-        .gw = {192, 168, 29, 1},                     // Gateway
-        .dns = {8, 8, 8, 8},                         // DNS server
-        .dhcp = NETINFO_STATIC};
+    wiz_NetInfo net_info = {.mac  = {0x02, 0x08, 0xDC, 0x12, 0x34, 0x56}, // MAC address
+                            .ip   = {192, 168, 29, 11},                   // IP address
+                            .sn   = {255, 255, 255, 0},                   // Subnet mask
+                            .gw   = {192, 168, 29, 1},                    // Gateway
+                            .dns  = {8, 8, 8, 8},                         // DNS server
+                            .dhcp = NETINFO_STATIC};
 
     network_initialize(net_info);
     print_network_information(net_info);
@@ -74,7 +73,7 @@ static void run_udp_echo_server(void)
 {
     // Create a UDP socket bound to our chosen port. This socket will listen for
     // incoming UDP packets and can send responses back to the source address/port.
-    printf("Starting UDP echo on port %u...\r\n", UDP_PORT);
+    printf("Starting UDP echo on port %u...\r\n", (unsigned)UDP_PORT);
 
     // socket() allocates one of the W5500's 8 hardware sockets
     // Parameters: socket_number, protocol_mode, local_port, flags
@@ -91,9 +90,9 @@ static void run_udp_echo_server(void)
     printf("Socket opened successfully in blocking mode\r\n");
 
     // Allocate buffers for UDP packet data and sender information
-    uint8_t recv_buf[2048]; // Payload buffer (max UDP payload in W5500)
-    uint8_t remote_ip[4];   // Remote (sender) IP address
-    uint16_t remote_port;   // Remote (sender) port number
+    uint8_t  recv_buf[2048]; // Payload buffer (max UDP payload in W5500)
+    uint8_t  remote_ip[4];   // Remote (sender) IP address
+    uint16_t remote_port;    // Remote (sender) port number
 
     printf("Entering main echo loop (blocking mode)...\r\n");
 
@@ -108,12 +107,13 @@ static void run_udp_echo_server(void)
         if (recv_len > 0)
         {
             // Packet received - log the sender and byte count
-            printf("RX %ld bytes from %u.%u.%u.%u:%u\r\n",
-                   (long)recv_len, remote_ip[0], remote_ip[1], remote_ip[2], remote_ip[3], remote_port);
+            printf("RX %ld bytes from %u.%u.%u.%u:%u\r\n", (long)recv_len, remote_ip[0],
+                   remote_ip[1], remote_ip[2], remote_ip[3], remote_port);
 
             // Echo the packet back to the sender
             // Parameters: socket_number, data_buffer, data_length, dest_ip, dest_port
-            int32_t sent_len = sendto(UDP_SOCK, recv_buf, (uint16_t)recv_len, remote_ip, remote_port);
+            int32_t sent_len =
+                sendto(UDP_SOCK, recv_buf, (uint16_t)recv_len, remote_ip, remote_port);
             if (sent_len < 0)
             {
                 printf("sendto() error %ld\r\n", (long)sent_len);
